+++ CHANGELOG.md
@@ -16,7 +16,7 @@ Versions marked with a number and date (e.g. Falcon Client v0.1.0 (2018-10-05)) 
 
 ### Falcon Magento2 API vNext
 
-- added oAuth authorization for integration requests ([#424](https://github.com/deity-io/falcon/pull/424))
+- added oAuth authorization type for integration requests ([#424](https://github.com/deity-io/falcon/pull/424))
 
 ### Falcon E-commerce UI Kit vNext